+++ lionengine-game/src/main/java/com/b3dgs/lionengine/game/map/astar/HeuristicClosest.java
@@ -39,8 +39,8 @@ public final class HeuristicClosest implements Heuristic
     @Override
     public double getCost(int sx, int sy, int dx, int dy)
     {
-        final double x = dx - sx;
-        final double y = dy - sy;
+        final double x = dx - (double) sx;
+        final double y = dy - (double) sy;
         return Math.sqrt(x * x + y * y);
     }
 }
